Added Universe Import Version

+++ trunk/styles/templates/adm/UniversePage.tpl
@@ -1,4 +1,5 @@
 {include file="adm/overall_header.tpl"}
+<form enctype="multipart/form-data" action="?page=universe&action=import&sid={$SID}" method="POST">
 <table>
 	<tr>
 		<td class="c">ID</td>
@@ -17,10 +18,12 @@
 		<th>{$UniInfo.users_amount}</th>
 		<th>{if $UniInfo.game_disable == 0}N{else}Y{/if}</th>
 		<th><a href="?page=universe&action=download&id={$ID}&sid={$SID}"><img src="./styles/images/adm/Go.png" alt=""></a></th>
-		<th><a href="?page=universe&action=delete&id={$ID}&sid={$SID}"><img src="./styles/images/adm/i.gif" alt=""></a></th>
+		<th><a href="?page=universe&action=delete&id={$ID}&sid={$SID}" onclick="return confirm('Delete?');"><img src="./styles/images/adm/i.gif" alt=""></a></th>
 	</tr>
 	{/foreach}
-	<tr>
-	<th colspan="7"><a href="?page=universe&action=create&sid={$SID}">Create a new</a></th>
+	<tr><th colspan="7"><a href="?page=universe&action=create&sid={$SID}">Create a new</a></th></tr>
+	<tr><td class="c" colspan="7">Import a Universe</td></tr>
+	<tr><th colspan="7"><input name="file" type="file"><br><input type="submit" value="Upload"></th></tr>
 </table>
+</form>
 {include file="adm/overall_footer.tpl"}
